+++ lib/src/dev_utils.dart
@@ -3,9 +3,17 @@ bool _devPrintEnabled = true;
 @deprecated
 set devPrintEnabled(bool enabled) => _devPrintEnabled = enabled;
 
+/// Deprecated to prevent keeping the code used.
 @deprecated
 void devPrint(Object object) {
   if (_devPrintEnabled) {
     print(object);
   }
 }
+
+/// Deprecated to prevent keeping the code used.
+///
+/// Can be use as a todo for weird code. int value = devWarning(myFunction());
+/// The function is always called
+@deprecated
+T devWarning<T>(T value) => value;