+++ test/database_test.dart
@@ -83,7 +83,10 @@ void defineTests(DatabaseTestContext ctx) {
         var db = await factory.openDatabase(dbPath);
         try {
           // don't await to make sure it gets written at some point
-          unawaited(db.put('test', 1));
+          unawaited(db.transaction((txn) async {
+            await Future.delayed(const Duration(milliseconds: 10));
+            await txn.put('test', 1);
+          }));
           unawaited(db.close());
           db = await factory.openDatabase(dbPath);
           expect(await db.get(1), 'test');